requirements.txt
tensorflow>=2.15
keras>=2.15
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.8
seaborn>=0.12
scikit-learn>=1.3
pandas>=2.1